+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 -   Updated the following Components
 
+    -   Display
+
+        -   `List`
+
+            -   Updated to be in line with the rest of element configurable Components.
+
+                -   `<List.Ordered>` -> `<List.Container is="ol">`
+                -   `<List.Unordered>` -> `<List.Container is="ul">`
+
     -   Layouts
 
         -   `Grid`
@@ -17,6 +26,13 @@
             -   Added `Grid.Item` — Optionally used to wrap a child item.
 
                 -   `<Grid.Item span="1...12" span_x="1...12" span_y="1...12">` — Used to control how many Grid Points the child item spans in columns, rows, or both. With Viewport support, e.g. `<Grid.Item span={[3, "mobile:1"]}>`
+
+    -   Overlays
+
+        -   `Overlay`
+
+            -   Added `<Overlay on:active>` — Fires whenever the `Overlay` is made active, either via scripting or the end-user.
+            -   Added `<Overlay on:dismiss>` — Fires whenever the `Overlay` is dismissed, either via scripting or the end-user.
 
 ## v0.2.9 - 2021/06/24